Preparing for Alpha launch.
I modified some functions quite heavily, so hard to test core loop as it was not working. Now it does with the Usage_Example. I need to validate the order of the evaluate_design dataframes as they might be different after being processed, which could lead to a mismatch - very bad! So I have to check. I'll do that at a later stage though.